Fix connections not closing. Clean up environment variables.
There were issues with mongoose not closing connections when the tests
ran as well as the http server itself not shutting down properly. The
mongoose driver now disconnects at the end of the tests. Supertest
already handles shutting down the app for you, just had to factor out
the `app.listen` into a different file.